Conflicts:
	configure.ac
	src/core/system_decl.hpp

+++ src/core/cachedaccelerator_decl.hpp
@@ -65,7 +65,7 @@ namespace nanos
 
          unsigned int getMemorySpaceId() const { return _cache->getId(); }
 
-         void configureCache( int cacheSize, System::CachePolicyType cachePolicy );
+         void configureCache( size_t cacheSize, System::CachePolicyType cachePolicy );
 
          void registerCacheAccessDependent( Directory& dir, uint64_t tag, size_t size, bool input, bool output );
